feat: unify database and sync status indicators into a single combined chip
